增加了一个闪屏页，增加了一个Activity管理类，BaseActivity中增加了将Activity从管理类添加和移除的方法，管理类中有直接finish掉所有activity的方法，教师包下的activity均增加了一个静态方法actionStart()，其它类进行跳转时可调用。

+++ LeXueBDemo/app/src/main/java/com/example/dllo/lexuebdemo/teacher/model/CommentBean.java
@@ -18,7 +18,7 @@ public class CommentBean {
      */
 
     private int comment_display;
-    private int comment_rate;
+    private double comment_rate;
     private String error_info;
     private int status;
     private int total;
@@ -32,13 +32,13 @@ public class CommentBean {
         this.comment_display = comment_display;
     }
 
-    public int getComment_rate() {
+    public double getComment_rate() {
         return comment_rate;
     }
-
-    public void setComment_rate(int comment_rate) {
+    public void setComment_rate(double comment_rate){
         this.comment_rate = comment_rate;
     }
+
 
     public String getError_info() {
         return error_info;
